pyproject.toml
[build-system]
requires = ["hatchling", "hatch-zipped-directory"]
build-backend = "hatchling.build"

[project]
name = "sbijax"
description = " Simulation-based inference in JAX"
authors = [{name = "Simon Dirmeier", email = "sfyrbnd@pm.me"}]
readme = "README.md"
license = "Apache-2.0"
keywords = ["sbi", "abc", "simulation-based inference", "approximate Bayesian computation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "arviz>=0.17.1",
    "diffrax>=0.6.1",
    "blackjax-nightly>=1.0.0.post17",
    "dm-haiku>=0.0.9",
    "einops>=0.8.0",
    "matplotlib>=3.6.2",
    "optax>=0.1.3",
    "seaborn>=0.12.2",
    "surjectors>=0.3.3",
    "tensorflow>=2.15.0",
    "tensorflow-datasets>=4.9.3",
    "tfp-nightly>=0.20.0.dev20230404",
    "tqdm>=4.64.1"
]
dynamic = ["version"]

[project.optional-dependencies]
examples = [
  "jrnmm==0.1.1.post2",
]
[project.urls]
Homepage = "https://github.com/dirmeier/sbijax"
Documentation = "https://sbijax.readthedocs.io"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "sbijax/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "./gitignore",
    "/.pre-commit-config.yaml",
    "manuscript+supplement",
    "docs"
]

[tool.hatch.build.targets.zipped-directory]
exclude = [
    "/.gitattributes",
    "/.gitignore",
    "/.gitlint",
    "/.github",
    "./gitignore",
    "/.pre-commit-config.yaml",
    "/.readthedocs.yaml",
    "/Makefile",
    "/examples",
    "/manuscript+supplement",
]

[tool.hatch.envs.test]
dependencies = [
    "ruff>=0.3.0",
    "pytest>=7.2.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.envs.test.scripts]
lint = 'ruff check sbijax examples'
format = 'ruff format sbijax examples'
test = 'pytest -v --doctest-modules --cov=./sbijax --cov-report=xml sbijax'

[tool.bandit]
skips = ["B101"]

[tool.ruff]
line-length = 80
exclude = ["*_test.py", "docs/**", "examples/**", "manuscript+supplement/**"]

[tool.ruff.lint]
select = ["E", "F"]
extend-select = [
  "UP", "D", "I", "PL", "S"
]
ignore=["S101"]

[tool.ruff.lint.pydocstyle]
convention= 'google'
